Improves the error message
This adds some quotes around some data which would make it easy to spot
whitespace errors. Also, the error message that the given JCE provider
does not handle X.509 is now more vocal and tells the developer that
the following exception is not the end of the world and he did nothing
wrong at that point.

git-svn-id: https://svn.apache.org/repos/asf/webservices/wss4j/trunk@1732860 13f79535-47bb-0310-9956-ffa450edef68